requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.6
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.2.0
python-dotenv==1.0.1
Authlib==1.3.2
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0
httpx==0.27.2
